Project restructured, provided stats for landing
