+++ controllers/AdminController/BeeFamilyStatusController.go
@@ -1,4 +1,4 @@
-package CRUD
+package AdminController
 
 import (
 	"encoding/json"
@@ -13,9 +13,9 @@ import (
 	"strconv"
 )
 
-var NewsCreate = func(w http.ResponseWriter, r *http.Request) {
-	News := &models.News{}
-	err := json.NewDecoder(r.Body).Decode(News)
+var BeeFamilyStatusCreate = func(w http.ResponseWriter, r *http.Request) {
+	BeeFamilyStatus := &models.BeeFamilyStatus{}
+	err := json.NewDecoder(r.Body).Decode(BeeFamilyStatus)
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
@@ -23,24 +23,24 @@ var NewsCreate = func(w http.ResponseWriter, r *http.Request) {
 	}
 
 	db := db.GetDB()
-	err = db.Create(News).Error
+	err = db.Create(BeeFamilyStatus).Error
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
 	} else {
-		res, _ := json.Marshal(News)
+		res, _ := json.Marshal(BeeFamilyStatus)
 		u.RespondJSON(w, res)
 	}
 }
 
-var NewsRetrieve = func(w http.ResponseWriter, r *http.Request) {
-	News := &models.News{}
+var BeeFamilyStatusRetrieve = func(w http.ResponseWriter, r *http.Request) {
+	BeeFamilyStatus := &models.BeeFamilyStatus{}
 
 	params := mux.Vars(r)
 	id := params["id"]
 
 	db := db.GetDB()
-	err := db.First(&News, id).Error
+	err := db.First(&BeeFamilyStatus, id).Error
 
 	if err != nil {
 		if err == gorm.ErrRecordNotFound {
@@ -51,24 +51,24 @@ var NewsRetrieve = func(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	res, err := json.Marshal(News)
+	res, err := json.Marshal(BeeFamilyStatus)
 	if err != nil {
 		u.HandleBadRequest(w, err)
-	} else if News.ID == 0 {
+	} else if BeeFamilyStatus.ID == 0 {
 		u.HandleNotFound(w)
 	} else {
 		u.RespondJSON(w, res)
 	}
 }
 
-var NewsUpdate = func(w http.ResponseWriter, r *http.Request) {
-	News := &models.News{}
+var BeeFamilyStatusUpdate = func(w http.ResponseWriter, r *http.Request) {
+	BeeFamilyStatus := &models.BeeFamilyStatus{}
 
 	params := mux.Vars(r)
 	id := params["id"]
 
 	db := db.GetDB()
-	err := db.First(&News, id).Error
+	err := db.First(&BeeFamilyStatus, id).Error
 
 	if err != nil {
 		if err == gorm.ErrRecordNotFound {
@@ -79,15 +79,15 @@ var NewsUpdate = func(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	newNews := &models.News{}
-	err = json.NewDecoder(r.Body).Decode(newNews)
+	newBeeFamilyStatus := &models.BeeFamilyStatus{}
+	err = json.NewDecoder(r.Body).Decode(newBeeFamilyStatus)
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
 		return
 	}
 
-	err = db.Model(&News).Updates(newNews).Error
+	err = db.Model(&BeeFamilyStatus).Updates(newBeeFamilyStatus).Error
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
@@ -96,12 +96,12 @@ var NewsUpdate = func(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-var NewsDelete = func(w http.ResponseWriter, r *http.Request) {
+var BeeFamilyStatusDelete = func(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	id := params["id"]
 
 	db := db.GetDB()
-	err := db.Delete(&models.News{}, id).Error
+	err := db.Delete(&models.BeeFamilyStatus{}, id).Error
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
@@ -110,8 +110,8 @@ var NewsDelete = func(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-var NewsQuery = func(w http.ResponseWriter, r *http.Request) {
-	var entities []models.News
+var BeeFamilyStatusQuery = func(w http.ResponseWriter, r *http.Request) {
+	var entities []models.BeeFamilyStatus
 	var count string
 
 	order := r.FormValue("_order")
@@ -138,7 +138,7 @@ var NewsQuery = func(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		u.HandleBadRequest(w, err)
 	} else {
-		db.Model(&models.News{}).Count(&count)
+		db.Model(&models.BeeFamilyStatus{}).Count(&count)
 		u.SetTotalCountHeader(w, count)
 		u.RespondJSON(w, res)
 	}
