+++ controllers/AdminController/HiveFrameTypeController.go
@@ -1,4 +1,4 @@
-package CRUD
+package AdminController
 
 import (
 	"encoding/json"
@@ -13,9 +13,9 @@ import (
 	"strconv"
 )
 
-var HoneyHarvestCreate = func(w http.ResponseWriter, r *http.Request) {
-	HoneyHarvest := &models.HoneyHarvest{}
-	err := json.NewDecoder(r.Body).Decode(HoneyHarvest)
+var HiveFrameTypeCreate = func(w http.ResponseWriter, r *http.Request) {
+	HiveFrameType := &models.HiveFrameType{}
+	err := json.NewDecoder(r.Body).Decode(HiveFrameType)
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
@@ -23,24 +23,24 @@ var HoneyHarvestCreate = func(w http.ResponseWriter, r *http.Request) {
 	}
 
 	db := db.GetDB()
-	err = db.Create(HoneyHarvest).Error
+	err = db.Create(HiveFrameType).Error
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
 	} else {
-		res, _ := json.Marshal(HoneyHarvest)
+		res, _ := json.Marshal(HiveFrameType)
 		u.RespondJSON(w, res)
 	}
 }
 
-var HoneyHarvestRetrieve = func(w http.ResponseWriter, r *http.Request) {
-	HoneyHarvest := &models.HoneyHarvest{}
+var HiveFrameTypeRetrieve = func(w http.ResponseWriter, r *http.Request) {
+	HiveFrameType := &models.HiveFrameType{}
 
 	params := mux.Vars(r)
 	id := params["id"]
 
 	db := db.GetDB()
-	err := db.Preload("HoneyType").Preload("BeeFamily").First(&HoneyHarvest, id).Error
+	err := db.First(&HiveFrameType, id).Error
 
 	if err != nil {
 		if err == gorm.ErrRecordNotFound {
@@ -51,24 +51,24 @@ var HoneyHarvestRetrieve = func(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	res, err := json.Marshal(HoneyHarvest)
+	res, err := json.Marshal(HiveFrameType)
 	if err != nil {
 		u.HandleBadRequest(w, err)
-	} else if HoneyHarvest.ID == 0 {
+	} else if HiveFrameType.ID == 0 {
 		u.HandleNotFound(w)
 	} else {
 		u.RespondJSON(w, res)
 	}
 }
 
-var HoneyHarvestUpdate = func(w http.ResponseWriter, r *http.Request) {
-	HoneyHarvest := &models.HoneyHarvest{}
+var HiveFrameTypeUpdate = func(w http.ResponseWriter, r *http.Request) {
+	HiveFrameType := &models.HiveFrameType{}
 
 	params := mux.Vars(r)
 	id := params["id"]
 
 	db := db.GetDB()
-	err := db.First(&HoneyHarvest, id).Error
+	err := db.First(&HiveFrameType, id).Error
 
 	if err != nil {
 		if err == gorm.ErrRecordNotFound {
@@ -79,17 +79,15 @@ var HoneyHarvestUpdate = func(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	newHoneyHarvest := &models.HoneyHarvest{}
-	err = json.NewDecoder(r.Body).Decode(newHoneyHarvest)
+	newHiveFrameType := &models.HiveFrameType{}
+	err = json.NewDecoder(r.Body).Decode(newHiveFrameType)
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
 		return
 	}
 
-	newHoneyHarvest.HoneyType = HoneyHarvest.HoneyType
-
-	err = db.Model(&HoneyHarvest).Updates(newHoneyHarvest).Error
+	err = db.Model(&HiveFrameType).Updates(newHiveFrameType).Error
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
@@ -98,12 +96,12 @@ var HoneyHarvestUpdate = func(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-var HoneyHarvestDelete = func(w http.ResponseWriter, r *http.Request) {
+var HiveFrameTypeDelete = func(w http.ResponseWriter, r *http.Request) {
 	params := mux.Vars(r)
 	id := params["id"]
 
 	db := db.GetDB()
-	err := db.Delete(&models.HoneyHarvest{}, id).Error
+	err := db.Delete(&models.HiveFrameType{}, id).Error
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
@@ -112,8 +110,8 @@ var HoneyHarvestDelete = func(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-var HoneyHarvestQuery = func(w http.ResponseWriter, r *http.Request) {
-	var entities []models.HoneyHarvest
+var HiveFrameTypeQuery = func(w http.ResponseWriter, r *http.Request) {
+	var entities []models.HiveFrameType
 	var count string
 
 	order := r.FormValue("_order")
@@ -128,8 +126,7 @@ var HoneyHarvestQuery = func(w http.ResponseWriter, r *http.Request) {
 	u.CheckOrderAndSortParams(&order, &sort)
 
 	db := db.GetDB()
-	err := db.Preload("HoneyType").Preload("BeeFamily").
-		Order(fmt.Sprintf("%s %s", sort, order)).Offset(start).Limit(end + start).Find(&entities).Error
+	err := db.Order(fmt.Sprintf("%s %s", sort, order)).Offset(start).Limit(end + start).Find(&entities).Error
 
 	if err != nil {
 		u.HandleBadRequest(w, err)
@@ -141,7 +138,7 @@ var HoneyHarvestQuery = func(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		u.HandleBadRequest(w, err)
 	} else {
-		db.Model(&models.HoneyHarvest{}).Count(&count)
+		db.Model(&models.HiveFrameType{}).Count(&count)
 		u.SetTotalCountHeader(w, count)
 		u.RespondJSON(w, res)
 	}
