+++ controllers/AdminController/SubscriptionTypeController.go
@@ -1,4 +1,4 @@
-package CRUD
+package AdminController
 
 import (
 	"encoding/json"
