+++ controllers/MainController/ControllersLK/SubscriptionTypeController.go
@@ -1,4 +1,4 @@
-package OtherControllers
+package ControllersLK
 
 import (
 	"encoding/json"
