+++ controllers/MainController/CustomTypesControllers/BeeDiseaseController.go
