+++ router/router.go
@@ -2,13 +2,14 @@ package router
 
 import (
 	"github.com/gorilla/mux"
+	"paseca/controllers/AdminController"
 	"paseca/controllers/AuthController"
-	"paseca/controllers/CRUD"
-	"paseca/controllers/ControllerLK/BeeFarmControllers"
-	"paseca/controllers/ControllerLK/HarvestControllers"
-	"paseca/controllers/ControllerLK/CustomTypesControllers"
-	"paseca/controllers/ControllerLK/OtherControllers"
-	"paseca/controllers/ControllerLK/WikiControllers"
+	"paseca/controllers/MainController/BeeFarmControllers"
+	"paseca/controllers/MainController/ControllersLK"
+	"paseca/controllers/MainController/CustomTypesControllers"
+	"paseca/controllers/MainController/HarvestControllers"
+	"paseca/controllers/MainController/OtherControllers"
+	"paseca/controllers/MainController/WikiControllers"
 	"paseca/middleware"
 	u "paseca/utils"
 )
@@ -20,9 +21,10 @@ func InitRouter() *mux.Router {
 	api   := router.PathPrefix("/api").Subrouter()
 
 	// Subrouters
-	admin := api.PathPrefix("/admin").Subrouter()
-	auth  := api.PathPrefix("/auth").Subrouter()
-	lk    := api.PathPrefix("/lk").Subrouter()
+	admin   := api.PathPrefix("/admin").Subrouter()
+	auth    := api.PathPrefix("/auth").Subrouter()
+	lk      := api.PathPrefix("/lk").Subrouter()
+	landing := api.PathPrefix("/landing").Subrouter()
 
 	// Handle options
 	api.HandleFunc("/{any}", u.HandleOptions).Methods("OPTIONS")
@@ -30,156 +32,156 @@ func InitRouter() *mux.Router {
 	api.HandleFunc("/{any1}/{any2}/{any3}", u.HandleOptions).Methods("OPTIONS")
 	api.HandleFunc("/{any1}/{any2}/{any3}/{any4}", u.HandleOptions).Methods("OPTIONS")
 
-	// ADMIN CRUD
-	admin.HandleFunc("/bee_breed", CRUD.BeeBreedQuery).Methods("GET")
-	admin.HandleFunc("/bee_breed", CRUD.BeeBreedCreate).Methods("POST")
-	admin.HandleFunc("/bee_breed/{id}", CRUD.BeeBreedRetrieve).Methods("GET")
-	admin.HandleFunc("/bee_breed/{id}", CRUD.BeeBreedUpdate).Methods("PUT")
-	admin.HandleFunc("/bee_breed/{id}", CRUD.BeeBreedDelete).Methods("DELETE")
+	// ADMIN AdminController
+	admin.HandleFunc("/bee_breed", AdminController.BeeBreedQuery).Methods("GET")
+	admin.HandleFunc("/bee_breed", AdminController.BeeBreedCreate).Methods("POST")
+	admin.HandleFunc("/bee_breed/{id}", AdminController.BeeBreedRetrieve).Methods("GET")
+	admin.HandleFunc("/bee_breed/{id}", AdminController.BeeBreedUpdate).Methods("PUT")
+	admin.HandleFunc("/bee_breed/{id}", AdminController.BeeBreedDelete).Methods("DELETE")
 
-	admin.HandleFunc("/bee_disease", CRUD.BeeDiseaseQuery).Methods("GET")
-	admin.HandleFunc("/bee_disease", CRUD.BeeDiseaseCreate).Methods("POST")
-	admin.HandleFunc("/bee_disease/{id}", CRUD.BeeDiseaseRetrieve).Methods("GET")
-	admin.HandleFunc("/bee_disease/{id}", CRUD.BeeDiseaseUpdate).Methods("PUT")
-	admin.HandleFunc("/bee_disease/{id}", CRUD.BeeDiseaseDelete).Methods("DELETE")
+	admin.HandleFunc("/bee_disease", AdminController.BeeDiseaseQuery).Methods("GET")
+	admin.HandleFunc("/bee_disease", AdminController.BeeDiseaseCreate).Methods("POST")
+	admin.HandleFunc("/bee_disease/{id}", AdminController.BeeDiseaseRetrieve).Methods("GET")
+	admin.HandleFunc("/bee_disease/{id}", AdminController.BeeDiseaseUpdate).Methods("PUT")
+	admin.HandleFunc("/bee_disease/{id}", AdminController.BeeDiseaseDelete).Methods("DELETE")
 
-	admin.HandleFunc("/bee_family", CRUD.BeeFamilyQuery).Methods("GET")
-	admin.HandleFunc("/bee_family", CRUD.BeeFamilyCreate).Methods("POST")
-	admin.HandleFunc("/bee_family/{id}", CRUD.BeeFamilyRetrieve).Methods("GET")
-	admin.HandleFunc("/bee_family/{id}", CRUD.BeeFamilyUpdate).Methods("PUT")
-	admin.HandleFunc("/bee_family/{id}", CRUD.BeeFamilyDelete).Methods("DELETE")
+	admin.HandleFunc("/bee_family", AdminController.BeeFamilyQuery).Methods("GET")
+	admin.HandleFunc("/bee_family", AdminController.BeeFamilyCreate).Methods("POST")
+	admin.HandleFunc("/bee_family/{id}", AdminController.BeeFamilyRetrieve).Methods("GET")
+	admin.HandleFunc("/bee_family/{id}", AdminController.BeeFamilyUpdate).Methods("PUT")
+	admin.HandleFunc("/bee_family/{id}", AdminController.BeeFamilyDelete).Methods("DELETE")
 
-	admin.HandleFunc("/bee_family_status", CRUD.BeeFamilyStatusQuery).Methods("GET")
-	admin.HandleFunc("/bee_family_status", CRUD.BeeFamilyStatusCreate).Methods("POST")
-	admin.HandleFunc("/bee_family_status/{id}", CRUD.BeeFamilyStatusRetrieve).Methods("GET")
-	admin.HandleFunc("/bee_family_status/{id}", CRUD.BeeFamilyStatusUpdate).Methods("PUT")
-	admin.HandleFunc("/bee_family_status/{id}", CRUD.BeeFamilyStatusDelete).Methods("DELETE")
+	admin.HandleFunc("/bee_family_status", AdminController.BeeFamilyStatusQuery).Methods("GET")
+	admin.HandleFunc("/bee_family_status", AdminController.BeeFamilyStatusCreate).Methods("POST")
+	admin.HandleFunc("/bee_family_status/{id}", AdminController.BeeFamilyStatusRetrieve).Methods("GET")
+	admin.HandleFunc("/bee_family_status/{id}", AdminController.BeeFamilyStatusUpdate).Methods("PUT")
+	admin.HandleFunc("/bee_family_status/{id}", AdminController.BeeFamilyStatusDelete).Methods("DELETE")
 
-	admin.HandleFunc("/bee_farm", CRUD.BeeFarmQuery).Methods("GET")
-	admin.HandleFunc("/bee_farm", CRUD.BeeFarmCreate).Methods("POST")
-	admin.HandleFunc("/bee_farm/{id}", CRUD.BeeFarmRetrieve).Methods("GET")
-	admin.HandleFunc("/bee_farm/{id}", CRUD.BeeFarmUpdate).Methods("PUT")
-	admin.HandleFunc("/bee_farm/{id}", CRUD.BeeFarmDelete).Methods("DELETE")
+	admin.HandleFunc("/bee_farm", AdminController.BeeFarmQuery).Methods("GET")
+	admin.HandleFunc("/bee_farm", AdminController.BeeFarmCreate).Methods("POST")
+	admin.HandleFunc("/bee_farm/{id}", AdminController.BeeFarmRetrieve).Methods("GET")
+	admin.HandleFunc("/bee_farm/{id}", AdminController.BeeFarmUpdate).Methods("PUT")
+	admin.HandleFunc("/bee_farm/{id}", AdminController.BeeFarmDelete).Methods("DELETE")
 
-	admin.HandleFunc("/bee_farm_size", CRUD.BeeFarmSizeQuery).Methods("GET")
-	admin.HandleFunc("/bee_farm_size", CRUD.BeeFarmSizeCreate).Methods("POST")
-	admin.HandleFunc("/bee_farm_size/{id}", CRUD.BeeFarmSizeRetrieve).Methods("GET")
-	admin.HandleFunc("/bee_farm_size/{id}", CRUD.BeeFarmSizeUpdate).Methods("PUT")
-	admin.HandleFunc("/bee_farm_size/{id}", CRUD.BeeFarmSizeDelete).Methods("DELETE")
+	admin.HandleFunc("/bee_farm_size", AdminController.BeeFarmSizeQuery).Methods("GET")
+	admin.HandleFunc("/bee_farm_size", AdminController.BeeFarmSizeCreate).Methods("POST")
+	admin.HandleFunc("/bee_farm_size/{id}", AdminController.BeeFarmSizeRetrieve).Methods("GET")
+	admin.HandleFunc("/bee_farm_size/{id}", AdminController.BeeFarmSizeUpdate).Methods("PUT")
+	admin.HandleFunc("/bee_farm_size/{id}", AdminController.BeeFarmSizeDelete).Methods("DELETE")
 
-	admin.HandleFunc("/bee_farm_type", CRUD.BeeFarmTypeQuery).Methods("GET")
-	admin.HandleFunc("/bee_farm_type", CRUD.BeeFarmTypeCreate).Methods("POST")
-	admin.HandleFunc("/bee_farm_type/{id}", CRUD.BeeFarmTypeRetrieve).Methods("GET")
-	admin.HandleFunc("/bee_farm_type/{id}", CRUD.BeeFarmTypeUpdate).Methods("PUT")
-	admin.HandleFunc("/bee_farm_type/{id}", CRUD.BeeFarmTypeDelete).Methods("DELETE")
+	admin.HandleFunc("/bee_farm_type", AdminController.BeeFarmTypeQuery).Methods("GET")
+	admin.HandleFunc("/bee_farm_type", AdminController.BeeFarmTypeCreate).Methods("POST")
+	admin.HandleFunc("/bee_farm_type/{id}", AdminController.BeeFarmTypeRetrieve).Methods("GET")
+	admin.HandleFunc("/bee_farm_type/{id}", AdminController.BeeFarmTypeUpdate).Methods("PUT")
+	admin.HandleFunc("/bee_farm_type/{id}", AdminController.BeeFarmTypeDelete).Methods("DELETE")
 
-	admin.HandleFunc("/control_harvest", CRUD.ControlHarvestQuery).Methods("GET")
-	admin.HandleFunc("/control_harvest", CRUD.ControlHarvestCreate).Methods("POST")
-	admin.HandleFunc("/control_harvest/{id}", CRUD.ControlHarvestRetrieve).Methods("GET")
-	admin.HandleFunc("/control_harvest/{id}", CRUD.ControlHarvestUpdate).Methods("PUT")
-	admin.HandleFunc("/control_harvest/{id}", CRUD.ControlHarvestDelete).Methods("DELETE")
+	admin.HandleFunc("/control_harvest", AdminController.ControlHarvestQuery).Methods("GET")
+	admin.HandleFunc("/control_harvest", AdminController.ControlHarvestCreate).Methods("POST")
+	admin.HandleFunc("/control_harvest/{id}", AdminController.ControlHarvestRetrieve).Methods("GET")
+	admin.HandleFunc("/control_harvest/{id}", AdminController.ControlHarvestUpdate).Methods("PUT")
+	admin.HandleFunc("/control_harvest/{id}", AdminController.ControlHarvestDelete).Methods("DELETE")
 
-	admin.HandleFunc("/family_disease", CRUD.FamilyDiseaseQuery).Methods("GET")
-	admin.HandleFunc("/family_disease", CRUD.FamilyDiseaseCreate).Methods("POST")
-	admin.HandleFunc("/family_disease/{id}", CRUD.FamilyDiseaseRetrieve).Methods("GET")
-	admin.HandleFunc("/family_disease/{id}", CRUD.FamilyDiseaseUpdate).Methods("PUT")
-	admin.HandleFunc("/family_disease/{id}", CRUD.FamilyDiseaseDelete).Methods("DELETE")
+	admin.HandleFunc("/family_disease", AdminController.FamilyDiseaseQuery).Methods("GET")
+	admin.HandleFunc("/family_disease", AdminController.FamilyDiseaseCreate).Methods("POST")
+	admin.HandleFunc("/family_disease/{id}", AdminController.FamilyDiseaseRetrieve).Methods("GET")
+	admin.HandleFunc("/family_disease/{id}", AdminController.FamilyDiseaseUpdate).Methods("PUT")
+	admin.HandleFunc("/family_disease/{id}", AdminController.FamilyDiseaseDelete).Methods("DELETE")
 
-	admin.HandleFunc("/hive", CRUD.HiveQuery).Methods("GET")
-	admin.HandleFunc("/hive", CRUD.HiveCreate).Methods("POST")
-	admin.HandleFunc("/hive/{id}", CRUD.HiveRetrieve).Methods("GET")
-	admin.HandleFunc("/hive/{id}", CRUD.HiveUpdate).Methods("PUT")
-	admin.HandleFunc("/hive/{id}", CRUD.HiveDelete).Methods("DELETE")
+	admin.HandleFunc("/hive", AdminController.HiveQuery).Methods("GET")
+	admin.HandleFunc("/hive", AdminController.HiveCreate).Methods("POST")
+	admin.HandleFunc("/hive/{id}", AdminController.HiveRetrieve).Methods("GET")
+	admin.HandleFunc("/hive/{id}", AdminController.HiveUpdate).Methods("PUT")
+	admin.HandleFunc("/hive/{id}", AdminController.HiveDelete).Methods("DELETE")
 
-	admin.HandleFunc("/hive_format", CRUD.HiveFormatQuery).Methods("GET")
-	admin.HandleFunc("/hive_format", CRUD.HiveFormatCreate).Methods("POST")
-	admin.HandleFunc("/hive_format/{id}", CRUD.HiveFormatRetrieve).Methods("GET")
-	admin.HandleFunc("/hive_format/{id}", CRUD.HiveFormatUpdate).Methods("PUT")
-	admin.HandleFunc("/hive_format/{id}", CRUD.HiveFormatDelete).Methods("DELETE")
+	admin.HandleFunc("/hive_format", AdminController.HiveFormatQuery).Methods("GET")
+	admin.HandleFunc("/hive_format", AdminController.HiveFormatCreate).Methods("POST")
+	admin.HandleFunc("/hive_format/{id}", AdminController.HiveFormatRetrieve).Methods("GET")
+	admin.HandleFunc("/hive_format/{id}", AdminController.HiveFormatUpdate).Methods("PUT")
+	admin.HandleFunc("/hive_format/{id}", AdminController.HiveFormatDelete).Methods("DELETE")
 
-	admin.HandleFunc("/hive_frame_type", CRUD.HiveFrameTypeQuery).Methods("GET")
-	admin.HandleFunc("/hive_frame_type", CRUD.HiveFrameTypeCreate).Methods("POST")
-	admin.HandleFunc("/hive_frame_type/{id}", CRUD.HiveFrameTypeRetrieve).Methods("GET")
-	admin.HandleFunc("/hive_frame_type/{id}", CRUD.HiveFrameTypeUpdate).Methods("PUT")
-	admin.HandleFunc("/hive_frame_type/{id}", CRUD.HiveFrameTypeDelete).Methods("DELETE")
+	admin.HandleFunc("/hive_frame_type", AdminController.HiveFrameTypeQuery).Methods("GET")
+	admin.HandleFunc("/hive_frame_type", AdminController.HiveFrameTypeCreate).Methods("POST")
+	admin.HandleFunc("/hive_frame_type/{id}", AdminController.HiveFrameTypeRetrieve).Methods("GET")
+	admin.HandleFunc("/hive_frame_type/{id}", AdminController.HiveFrameTypeUpdate).Methods("PUT")
+	admin.HandleFunc("/hive_frame_type/{id}", AdminController.HiveFrameTypeDelete).Methods("DELETE")
 
-	admin.HandleFunc("/honey_harvest", CRUD.HoneyHarvestQuery).Methods("GET")
-	admin.HandleFunc("/honey_harvest", CRUD.HoneyHarvestCreate).Methods("POST")
-	admin.HandleFunc("/honey_harvest/{id}", CRUD.HoneyHarvestRetrieve).Methods("GET")
-	admin.HandleFunc("/honey_harvest/{id}", CRUD.HoneyHarvestUpdate).Methods("PUT")
-	admin.HandleFunc("/honey_harvest/{id}", CRUD.HoneyHarvestDelete).Methods("DELETE")
+	admin.HandleFunc("/honey_harvest", AdminController.HoneyHarvestQuery).Methods("GET")
+	admin.HandleFunc("/honey_harvest", AdminController.HoneyHarvestCreate).Methods("POST")
+	admin.HandleFunc("/honey_harvest/{id}", AdminController.HoneyHarvestRetrieve).Methods("GET")
+	admin.HandleFunc("/honey_harvest/{id}", AdminController.HoneyHarvestUpdate).Methods("PUT")
+	admin.HandleFunc("/honey_harvest/{id}", AdminController.HoneyHarvestDelete).Methods("DELETE")
 
-	admin.HandleFunc("/honey_sale", CRUD.HoneySaleQuery).Methods("GET")
-	admin.HandleFunc("/honey_sale", CRUD.HoneySaleCreate).Methods("POST")
-	admin.HandleFunc("/honey_sale/{id}", CRUD.HoneySaleRetrieve).Methods("GET")
-	admin.HandleFunc("/honey_sale/{id}", CRUD.HoneySaleUpdate).Methods("PUT")
-	admin.HandleFunc("/honey_sale/{id}", CRUD.HoneySaleDelete).Methods("DELETE")
+	admin.HandleFunc("/honey_sale", AdminController.HoneySaleQuery).Methods("GET")
+	admin.HandleFunc("/honey_sale", AdminController.HoneySaleCreate).Methods("POST")
+	admin.HandleFunc("/honey_sale/{id}", AdminController.HoneySaleRetrieve).Methods("GET")
+	admin.HandleFunc("/honey_sale/{id}", AdminController.HoneySaleUpdate).Methods("PUT")
+	admin.HandleFunc("/honey_sale/{id}", AdminController.HoneySaleDelete).Methods("DELETE")
 
-	admin.HandleFunc("/honey_type", CRUD.HoneyTypeQuery).Methods("GET")
-	admin.HandleFunc("/honey_type", CRUD.HoneyTypeCreate).Methods("POST")
-	admin.HandleFunc("/honey_type/{id}", CRUD.HoneyTypeRetrieve).Methods("GET")
-	admin.HandleFunc("/honey_type/{id}", CRUD.HoneyTypeUpdate).Methods("PUT")
-	admin.HandleFunc("/honey_type/{id}", CRUD.HoneyTypeDelete).Methods("DELETE")
+	admin.HandleFunc("/honey_type", AdminController.HoneyTypeQuery).Methods("GET")
+	admin.HandleFunc("/honey_type", AdminController.HoneyTypeCreate).Methods("POST")
+	admin.HandleFunc("/honey_type/{id}", AdminController.HoneyTypeRetrieve).Methods("GET")
+	admin.HandleFunc("/honey_type/{id}", AdminController.HoneyTypeUpdate).Methods("PUT")
+	admin.HandleFunc("/honey_type/{id}", AdminController.HoneyTypeDelete).Methods("DELETE")
 
-	admin.HandleFunc("/reminder", CRUD.ReminderQuery).Methods("GET")
-	admin.HandleFunc("/reminder", CRUD.ReminderCreate).Methods("POST")
-	admin.HandleFunc("/reminder/{id}", CRUD.ReminderRetrieve).Methods("GET")
-	admin.HandleFunc("/reminder/{id}", CRUD.ReminderUpdate).Methods("PUT")
-	admin.HandleFunc("/reminder/{id}", CRUD.ReminderDelete).Methods("DELETE")
+	admin.HandleFunc("/reminder", AdminController.ReminderQuery).Methods("GET")
+	admin.HandleFunc("/reminder", AdminController.ReminderCreate).Methods("POST")
+	admin.HandleFunc("/reminder/{id}", AdminController.ReminderRetrieve).Methods("GET")
+	admin.HandleFunc("/reminder/{id}", AdminController.ReminderUpdate).Methods("PUT")
+	admin.HandleFunc("/reminder/{id}", AdminController.ReminderDelete).Methods("DELETE")
 
-	admin.HandleFunc("/news", CRUD.NewsQuery).Methods("GET")
-	admin.HandleFunc("/news", CRUD.NewsCreate).Methods("POST")
-	admin.HandleFunc("/news/{id}", CRUD.NewsRetrieve).Methods("GET")
-	admin.HandleFunc("/news/{id}", CRUD.NewsUpdate).Methods("PUT")
-	admin.HandleFunc("/news/{id}", CRUD.NewsDelete).Methods("DELETE")
+	admin.HandleFunc("/news", AdminController.NewsQuery).Methods("GET")
+	admin.HandleFunc("/news", AdminController.NewsCreate).Methods("POST")
+	admin.HandleFunc("/news/{id}", AdminController.NewsRetrieve).Methods("GET")
+	admin.HandleFunc("/news/{id}", AdminController.NewsUpdate).Methods("PUT")
+	admin.HandleFunc("/news/{id}", AdminController.NewsDelete).Methods("DELETE")
 
-	admin.HandleFunc("/pollen_harvest", CRUD.PollenHarvestQuery).Methods("GET")
-	admin.HandleFunc("/pollen_harvest", CRUD.PollenHarvestCreate).Methods("POST")
-	admin.HandleFunc("/pollen_harvest/{id}", CRUD.PollenHarvestRetrieve).Methods("GET")
-	admin.HandleFunc("/pollen_harvest/{id}", CRUD.PollenHarvestUpdate).Methods("PUT")
-	admin.HandleFunc("/pollen_harvest/{id}", CRUD.PollenHarvestDelete).Methods("DELETE")
+	admin.HandleFunc("/pollen_harvest", AdminController.PollenHarvestQuery).Methods("GET")
+	admin.HandleFunc("/pollen_harvest", AdminController.PollenHarvestCreate).Methods("POST")
+	admin.HandleFunc("/pollen_harvest/{id}", AdminController.PollenHarvestRetrieve).Methods("GET")
+	admin.HandleFunc("/pollen_harvest/{id}", AdminController.PollenHarvestUpdate).Methods("PUT")
+	admin.HandleFunc("/pollen_harvest/{id}", AdminController.PollenHarvestDelete).Methods("DELETE")
 
-	admin.HandleFunc("/subscription_status", CRUD.SubscriptionStatusQuery).Methods("GET")
-	admin.HandleFunc("/subscription_status", CRUD.SubscriptionStatusCreate).Methods("POST")
-	admin.HandleFunc("/subscription_status/{id}", CRUD.SubscriptionStatusRetrieve).Methods("GET")
-	admin.HandleFunc("/subscription_status/{id}", CRUD.SubscriptionStatusUpdate).Methods("PUT")
-	admin.HandleFunc("/subscription_status/{id}", CRUD.SubscriptionStatusDelete).Methods("DELETE")
+	admin.HandleFunc("/subscription_status", AdminController.SubscriptionStatusQuery).Methods("GET")
+	admin.HandleFunc("/subscription_status", AdminController.SubscriptionStatusCreate).Methods("POST")
+	admin.HandleFunc("/subscription_status/{id}", AdminController.SubscriptionStatusRetrieve).Methods("GET")
+	admin.HandleFunc("/subscription_status/{id}", AdminController.SubscriptionStatusUpdate).Methods("PUT")
+	admin.HandleFunc("/subscription_status/{id}", AdminController.SubscriptionStatusDelete).Methods("DELETE")
 
-	admin.HandleFunc("/subscription_type", CRUD.SubscriptionTypeQuery).Methods("GET")
-	admin.HandleFunc("/subscription_type", CRUD.SubscriptionTypeCreate).Methods("POST")
-	admin.HandleFunc("/subscription_type/{id}", CRUD.SubscriptionTypeRetrieve).Methods("GET")
-	admin.HandleFunc("/subscription_type/{id}", CRUD.SubscriptionTypeUpdate).Methods("PUT")
-	admin.HandleFunc("/subscription_type/{id}", CRUD.SubscriptionTypeDelete).Methods("DELETE")
+	admin.HandleFunc("/subscription_type", AdminController.SubscriptionTypeQuery).Methods("GET")
+	admin.HandleFunc("/subscription_type", AdminController.SubscriptionTypeCreate).Methods("POST")
+	admin.HandleFunc("/subscription_type/{id}", AdminController.SubscriptionTypeRetrieve).Methods("GET")
+	admin.HandleFunc("/subscription_type/{id}", AdminController.SubscriptionTypeUpdate).Methods("PUT")
+	admin.HandleFunc("/subscription_type/{id}", AdminController.SubscriptionTypeDelete).Methods("DELETE")
 
-	admin.HandleFunc("/swarm", CRUD.SwarmQuery).Methods("GET")
-	admin.HandleFunc("/swarm", CRUD.SwarmCreate).Methods("POST")
-	admin.HandleFunc("/swarm/{id}", CRUD.SwarmRetrieve).Methods("GET")
-	admin.HandleFunc("/swarm/{id}", CRUD.SwarmUpdate).Methods("PUT")
-	admin.HandleFunc("/swarm/{id}", CRUD.SwarmDelete).Methods("DELETE")
+	admin.HandleFunc("/swarm", AdminController.SwarmQuery).Methods("GET")
+	admin.HandleFunc("/swarm", AdminController.SwarmCreate).Methods("POST")
+	admin.HandleFunc("/swarm/{id}", AdminController.SwarmRetrieve).Methods("GET")
+	admin.HandleFunc("/swarm/{id}", AdminController.SwarmUpdate).Methods("PUT")
+	admin.HandleFunc("/swarm/{id}", AdminController.SwarmDelete).Methods("DELETE")
 
-	admin.HandleFunc("/swarm_status", CRUD.SwarmStatusQuery).Methods("GET")
-	admin.HandleFunc("/swarm_status", CRUD.SwarmStatusCreate).Methods("POST")
-	admin.HandleFunc("/swarm_status/{id}", CRUD.SwarmStatusRetrieve).Methods("GET")
-	admin.HandleFunc("/swarm_status/{id}", CRUD.SwarmStatusUpdate).Methods("PUT")
-	admin.HandleFunc("/swarm_status/{id}", CRUD.SwarmStatusDelete).Methods("DELETE")
+	admin.HandleFunc("/swarm_status", AdminController.SwarmStatusQuery).Methods("GET")
+	admin.HandleFunc("/swarm_status", AdminController.SwarmStatusCreate).Methods("POST")
+	admin.HandleFunc("/swarm_status/{id}", AdminController.SwarmStatusRetrieve).Methods("GET")
+	admin.HandleFunc("/swarm_status/{id}", AdminController.SwarmStatusUpdate).Methods("PUT")
+	admin.HandleFunc("/swarm_status/{id}", AdminController.SwarmStatusDelete).Methods("DELETE")
 
-	admin.HandleFunc("/user", CRUD.UserQuery).Methods("GET")
-	admin.HandleFunc("/user", CRUD.UserCreate).Methods("POST")
-	admin.HandleFunc("/user/{id}", CRUD.UserRetrieve).Methods("GET")
-	admin.HandleFunc("/user/{id}", CRUD.UserUpdate).Methods("PUT")
-	admin.HandleFunc("/user/{id}", CRUD.UserDelete).Methods("DELETE")
+	admin.HandleFunc("/user", AdminController.UserQuery).Methods("GET")
+	admin.HandleFunc("/user", AdminController.UserCreate).Methods("POST")
+	admin.HandleFunc("/user/{id}", AdminController.UserRetrieve).Methods("GET")
+	admin.HandleFunc("/user/{id}", AdminController.UserUpdate).Methods("PUT")
+	admin.HandleFunc("/user/{id}", AdminController.UserDelete).Methods("DELETE")
 
-	admin.HandleFunc("/wiki_page", CRUD.WikiPageQuery).Methods("GET")
-	admin.HandleFunc("/wiki_page", CRUD.WikiPageCreate).Methods("POST")
-	admin.HandleFunc("/wiki_page/{id}", CRUD.WikiPageRetrieve).Methods("GET")
-	admin.HandleFunc("/wiki_page/{id}", CRUD.WikiPageUpdate).Methods("PUT")
-	admin.HandleFunc("/wiki_page/{id}", CRUD.WikiPageDelete).Methods("DELETE")
+	admin.HandleFunc("/wiki_page", AdminController.WikiPageQuery).Methods("GET")
+	admin.HandleFunc("/wiki_page", AdminController.WikiPageCreate).Methods("POST")
+	admin.HandleFunc("/wiki_page/{id}", AdminController.WikiPageRetrieve).Methods("GET")
+	admin.HandleFunc("/wiki_page/{id}", AdminController.WikiPageUpdate).Methods("PUT")
+	admin.HandleFunc("/wiki_page/{id}", AdminController.WikiPageDelete).Methods("DELETE")
 
-	admin.HandleFunc("/wiki_section", CRUD.WikiSectionQuery).Methods("GET")
-	admin.HandleFunc("/wiki_section", CRUD.WikiSectionCreate).Methods("POST")
-	admin.HandleFunc("/wiki_section/{id}", CRUD.WikiSectionRetrieve).Methods("GET")
-	admin.HandleFunc("/wiki_section/{id}", CRUD.WikiSectionUpdate).Methods("PUT")
-	admin.HandleFunc("/wiki_section/{id}", CRUD.WikiSectionDelete).Methods("DELETE")
+	admin.HandleFunc("/wiki_section", AdminController.WikiSectionQuery).Methods("GET")
+	admin.HandleFunc("/wiki_section", AdminController.WikiSectionCreate).Methods("POST")
+	admin.HandleFunc("/wiki_section/{id}", AdminController.WikiSectionRetrieve).Methods("GET")
+	admin.HandleFunc("/wiki_section/{id}", AdminController.WikiSectionUpdate).Methods("PUT")
+	admin.HandleFunc("/wiki_section/{id}", AdminController.WikiSectionDelete).Methods("DELETE")
 
 	// AUTH PATHS
 	auth.HandleFunc("/login", AuthController.Authenticate).Methods("POST")
@@ -265,7 +267,7 @@ func InitRouter() *mux.Router {
 	lk.HandleFunc("/check_reminder/{id}", BeeFarmControllers.CheckReminder).Methods("POST")
 	lk.HandleFunc("/reminders_by_bee_farm_id/{id}", BeeFarmControllers.GetRemindersByBeeFarmID).Methods("GET")
 
-	lk.HandleFunc("/subscription_types", OtherControllers.GetSubscriptionTypes).Methods("GET")
+	lk.HandleFunc("/subscription_types", ControllersLK.GetSubscriptionTypes).Methods("GET")
 
 	lk.HandleFunc("/swarm", BeeFarmControllers.CreateSwarm).Methods("POST")
 	lk.HandleFunc("/swarm/{id}", BeeFarmControllers.DeleteSwarm).Methods("DELETE")
@@ -275,11 +277,13 @@ func InitRouter() *mux.Router {
 	lk.HandleFunc("/swarm_status", CustomTypesControllers.CreateSwarmStatus).Methods("POST")
 	lk.HandleFunc("/swarm_status/{id}", CustomTypesControllers.DeleteSwarmStatusByID).Methods("DELETE")
 
-	lk.HandleFunc("/user", OtherControllers.GetUser).Methods("GET")
+	lk.HandleFunc("/user", ControllersLK.GetUser).Methods("GET")
 
 	lk.HandleFunc("/wiki_sections", WikiControllers.GetWikiSections).Methods("GET")
 	lk.HandleFunc("/wiki_pages_by_section_id/{id}", WikiControllers.GetWikiPagesBySectionID).Methods("GET")
 	lk.HandleFunc("/wiki_page/{id}", WikiControllers.GetWikiPageByID).Methods("GET")
+
+	landing.HandleFunc("/get_stats", OtherControllers.GetStatsForLanding).Methods("GET")
 
 	// middleware usage
 	// do NOT modify the order
